+++ changelog.md
@@ -1,7 +1,13 @@
 Changelog
 =========
 
-1.0.0-dev
-------------------
+1.1.0 2016-10-28
+----------------
+
+- Lazy load League Glide Server
+- Fix base_url examples
+
+1.0.0 2016-10-27
+----------------
 
 - First release.